add serializer, naming and texting updates [skip ci pytest]
