+++ frontend/i18next.scanner/en/translation.json
@@ -191,7 +191,7 @@
   "Error while deleting content": "Error while deleting content",
   "Error while moving content": "Error while moving content",
   "Error while uploading file": "Error while uploading file",
-  "Error while loading notification list": "Error while loading notification list",
+  "Error while loading the notification list": "Error while loading the notification list",
   "Error while marking the notification as read": "Error while marking the notification as read",
   "Move this content": "Move this content",
   "Move this folder": "Move this folder",
