+++ frontend/i18next.scanner/pt/translation.json
@@ -187,7 +187,7 @@
   "Error while deleting content": "Erro ao eliminar conteúdo",
   "Error while moving content": "Erro ao mover o conteúdo",
   "Error while uploading file": "Erro ao subir o ficheiro",
-  "Error while loading notification list": "Erro ao fazer upload da lista de notificações",
+  "Error while loading the notification list": "Erro ao fazer upload da lista de notificações",
   "Error while marking the notification as read": "Erro ao definir a notificação como lida",
   "Move this content": "Mover este conteúdo",
   "Move this folder": "Mover esta pasta",
